move get_ignored
